requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai>=1.20.0,<2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
reportlab
gunicorn==22.0.0
httplib2==0.22.0
httpx[http2]==0.28.1
celery==5.4.0
redis==5.0.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
idna==3.10
sqlalchemy==2.0.31
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dateutil==2.9.0
